--- sed/loader/fel/__init__.py
@@ -1,11 +1,0 @@
-"""sed.loader.fel module easy access APIs
-"""
-from .buffer import BufferHandler
-from .dataframe import DataFrameCreator
-from .parquet import ParquetHandler
-
-__all__ = [
-    "BufferHandler",
-    "DataFrameCreator",
-    "ParquetHandler",
-]
